fix: SSB display, millis overflow, audio state, EIBI search, web auth, BLE sleep
- Fix SSB frequency display with negative BFO (signed overflow)
- Fix signed long millis() overflow after 24.8 days uptime
- Fix audioTempMute/audioMuteForce state corruption with squelch
- Fix EIBI binary search off-by-one and eibiPrev() unsigned underflow
- Add HTTP basic auth to /setconfig and /api/command POST endpoints
- Clamp utcoffset and theme POST parameters to valid ranges
- Replace strcpy with strlcpy in Station.cpp for bounds safety
- Replace pow() with integer lookup table in Menu.cpp
- Change menu NVS saves from SAVE_ALL to SAVE_SETTINGS
- Add WiFi reconnection watchdog on link loss
- Allow clock re-sync if NTP drift exceeds 10 seconds
- Call bleStop/bleInit around CPU light sleep to preserve BLE
- Add skip-duplicate guard for short menus in MenuDraw.cpp
- Remove unused elapsedButton/lastStrengthCheck variables
- Set default AP password to "atsmini"
- Add String::reserve(8192) in web memory page builder
- Add volatile to savingPrefsFlag in Storage.cpp
- Rename REMOTE_DIRECTION to REMOTE_DIRECTION_SHIFT

+++ ats-mini/Scheduler.cpp
@@ -7,10 +7,10 @@
 #include "Storage.h"
 #include "Scheduler.h"
 
-extern long elapsedRSSI;
-extern long lastRDSCheck;
-extern long lastScheduleCheck;
-extern long lastNTPCheck;
+extern uint32_t elapsedRSSI;
+extern uint32_t lastRDSCheck;
+extern uint32_t lastScheduleCheck;
+extern uint32_t lastNTPCheck;
 extern uint32_t background_timer;
 
 bool runScheduler(uint32_t currentTime)
